+++ packages/falcon-client/src/client.js
@@ -5,49 +5,40 @@ import { BrowserRouter } from 'react-router-dom';
 import { ApolloProvider } from 'react-apollo';
 import { loadableReady } from '@loadable/component';
 import { I18nProvider } from '@deity/falcon-i18n';
-import { ApolloClient, apolloStateToObject } from './service';
+import { apolloClientWeb, apolloStateToObject } from './service';
 import HtmlHead from './components/HtmlHead';
 import App, { clientApolloSchema } from './clientApp';
 import i18nFactory from './i18n/i18nClientFactory';
-import { register, unregisterAll } from './serviceWorker';
+import { configureServiceWorker } from './serviceWorker';
 
 // eslint-disable-next-line no-underscore-dangle
-const apolloInitialState = window.__APOLLO_STATE__ || {};
-const i18nextState = window.I18NEXT_STATE || {};
+const initialState = window.__APOLLO_STATE__ || {};
+const config = apolloStateToObject(initialState, '$ROOT_QUERY.config');
 
-const config = apolloStateToObject(apolloInitialState, '$ROOT_QUERY.config');
-const apolloClient = new ApolloClient({
-  isBrowser: true,
-  clientState: clientApolloSchema,
-  initialState: apolloInitialState,
-  apolloClientConfig: config.apolloClient
-});
+const { language } = window.I18NEXT_STATE || {};
+const i18nConfig = { ...config.i18n, lng: language };
+const renderApp = config.serverSideRendering ? hydrate : render;
 
-i18nFactory({ ...config.i18n, lng: i18nextState.language }).then(i18next => {
-  const markup = (
-    <ApolloProvider client={apolloClient}>
-      <I18nProvider i18n={i18next}>
-        <BrowserRouter>
-          <React.Fragment>
-            <HtmlHead htmlLang={i18nextState.language || config.i18n.lng} />
-            <App />
-          </React.Fragment>
-        </BrowserRouter>
-      </I18nProvider>
-    </ApolloProvider>
-  );
-  const renderApp = config.serverSideRendering ? hydrate : render;
+loadableReady()
+  .then(() => apolloClientWeb({ initialState, clientApolloSchema, apolloClientConfig: config.apolloClient }))
+  .then(apolloClient => i18nFactory(i18nConfig).then(i18next => ({ apolloClient, i18next })))
+  .then(({ apolloClient, i18next }) => {
+    const markup = (
+      <ApolloProvider client={apolloClient}>
+        <I18nProvider i18n={i18next}>
+          <BrowserRouter>
+            <React.Fragment>
+              <HtmlHead htmlLang={i18nConfig.lng} />
+              <App />
+            </React.Fragment>
+          </BrowserRouter>
+        </I18nProvider>
+      </ApolloProvider>
+    );
 
-  loadableReady(() => {
     renderApp(markup, document.getElementById('root'));
-  });
-
-  if (process.env.NODE_ENV === 'production') {
-    register('/sw.js');
-  } else {
-    unregisterAll();
-  }
-});
+  })
+  .then(() => configureServiceWorker());
 
 if (module.hot) {
   module.hot.accept();