+++ packages/falcon-ecommerce-uikit/src/index.ts
@@ -29,5 +29,6 @@ export * from './AccountRecovery';
 export * from './Errors';
 export * from './SortOrdersProvider';
 export * from './Search';
+export * from './OnlineStatus';
 export * from './EnsureTTI';
 export * from './BackendConfig';