+++ CHANGELOG.md
@@ -18,6 +18,10 @@ Versions marked with a number and date (e.g. Falcon Client v0.1.0 (2018-10-05)) 
 
 - added `SearchProvider` with `SearchContext` that handles basic filtering ([#309](https://github.com/deity-io/falcon/pull/309))
 
+### Falcon Client
+
+- fixed offline mode ([#266](https://github.com/deity-io/falcon/pull/266))
+
 ---
 
 ## Falcon v0.3 (2019-01-18)